+++ src/Values/ElementAttribute.php
@@ -45,6 +45,14 @@ enum ElementAttribute: string
     case HEIGHT = 'height';
 
     /**
+     * `loading` — Indicates how the browser should load the image.
+     *
+     * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/img#loading
+     * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/iframe#loading
+     */
+    case LOADING = 'loading';
+
+    /**
      * `href` — The URL of a linked resource.
      *
      * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#href