+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Bug #53: Update documentation for various HTML attributes and their usage in elements (@terabytesoftw)
 - Enh #54: Add `HasDownload` trait and `download()` method to manage `download` attribute for HTML elements (@terabytesoftw)
 - Enh #55: Add `HasPing` trait and `ping()` method to manage `ping` attribute for HTML elements (@terabytesoftw)
+- Enh #56: Add `HasLoading` trait and `loading()` method to manage `loading` attribute for HTML elements (@terabytesoftw)
 
 ## 0.5.2 January 29, 2026